complete project push

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use App\Models\listing;
 use Illuminate\Foundation\Console\AboutCommand;
 use Illuminate\Support\Facades\Route;
@@ -21,21 +22,48 @@ use Illuminate\Support\Facades\Route;
 Route::get('/',[ListingController::class, 'index']);
 
 // Update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
-
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+  
 // Edit listing
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);     
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');     
 // show create listing
-Route::get('listing/create',  [ListingController::class, 'create']);
+Route::get('listing/create',  [ListingController::class, 'create'])->middleware('auth');
 
 // store listing
-Route::post('listings', [ListingController::class, 'store']);
+Route::post('listings', [ListingController::class, 'store'])->middleware('auth');
 
 // single listing
 Route::get('/listing/{listing}', [ListingController::class, 'show']);
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
+
+
+
+
+
+
+// User Registration
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// User logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Create new User
+Route::post('/users', [UserController::class,  'store']);
+
+// Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//  login
+Route::post('/users/authenticate', [UserController::class, 'Authenticate'])->name('Authenticate');
+
+
+
+
